requirements.txt
streamlit==1.36.0
tensorflow==2.16.1
torch==2.3.0
ultralytics==8.2.33
Pillow==10.3.0
numpy==1.26.4